[update] !!! 加入 hibiken/asynq 库依赖，基于 Redis 实现异步的处理架构。 1. main.go 中启动异步服务，注册异步处理过程枚举和处理 worker 方法； 2. 任务相关模型放在 models/task.go 中，internal/tasks/task.go 中，统一创建任务的方法； 3. woker 中 project.go 实现实际的项目相关异步任务处理过程，目前主要是删除前备份、下载前压缩； 4. 在 routes.go 中创建异步任务客户端，有需要的服务可引用此服务，添加异步任务； 5. 对前端暴露 /api/v1/tasks/{id} 的方法，用来轮询任务状态； 6. config.go 中新增异步的线程数量配置，目前默认为10； 7. 项目文件下载过程，已替换为 Asynq 异步处理方式；(前端还需要处理)
8. 删除 FileUtils 和 ZipUtils 等公共方法类，直接改用全局静态方法。新增 time_utils.go

+++ backend/internal/repositories/message_repository.go
@@ -25,7 +25,7 @@ type MessageRepository interface {
 	Delete(ctx context.Context, id string) error
 
 	// CountByProjectID 统计项目的对话消息数量
-	CountByProjectID(ctx context.Context, projectID string) (int64, error)
+	CountByProjectID(ctx context.Context, projectID string) (int, error)
 }
 
 // messageRepository 对话消息仓库实现
@@ -70,11 +70,11 @@ func (r *messageRepository) Delete(ctx context.Context, id string) error {
 	return r.db.WithContext(ctx).Delete(&models.ConversationMessage{}, "id = ?", id).Error
 }
 
-func (r *messageRepository) CountByProjectID(ctx context.Context, projectID string) (int64, error) {
+func (r *messageRepository) CountByProjectID(ctx context.Context, projectID string) (int, error) {
 	var count int64
 	err := r.db.WithContext(ctx).
 		Model(&models.ConversationMessage{}).
 		Where("project_id = ?", projectID).
 		Count(&count).Error
-	return count, err
+	return int(count), err
 }
